Moved some internal content to the internal directory
In preparation for some upcoming refactoring.
No functional/test changes.
