+++ client.go
@@ -2,11 +2,8 @@ package amqp
 
 import (
 	"context"
-	"encoding/base64"
 	"fmt"
-	"math/rand"
 	"net"
-	"sync"
 	"time"
 
 	"github.com/Azure/go-amqp/internal/debug"
@@ -150,32 +147,6 @@ type SessionOptions struct {
 	// Minimum: 1.
 	// Default: 4294967295.
 	MaxLinks uint32
-}
-
-// lockedRand provides a rand source that is safe for concurrent use.
-type lockedRand struct {
-	mu  sync.Mutex
-	src *rand.Rand
-}
-
-func (r *lockedRand) Read(p []byte) (int, error) {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	return r.src.Read(p)
-}
-
-// package scoped rand source to avoid any issues with seeding
-// of the global source.
-var pkgRand = &lockedRand{
-	src: rand.New(rand.NewSource(time.Now().UnixNano())),
-}
-
-// randBytes returns a base64 encoded string of n bytes.
-func randString(n int) string {
-	b := make([]byte, n)
-	// from math/rand, cannot fail
-	_, _ = pkgRand.Read(b)
-	return base64.RawURLEncoding.EncodeToString(b)
 }
 
 // linkKey uniquely identifies a link on a connection by name and direction.
